Target (cmd, api, build, pkg) renamed to TAP/TAPA
- cmd/tapa
- pkg/ambassador/tap
- api/ambassador
- target example tapa container
