+++ cmd/tapa/main.go
@@ -24,10 +24,10 @@ import (
 
 	"github.com/kelseyhightower/envconfig"
 	"github.com/networkservicemesh/sdk/pkg/tools/log"
+	"github.com/nordix/meridio/pkg/ambassador/tap"
 	"github.com/nordix/meridio/pkg/health"
 	linuxKernel "github.com/nordix/meridio/pkg/kernel"
 	"github.com/nordix/meridio/pkg/nsm"
-	"github.com/nordix/meridio/pkg/target"
 	"github.com/sirupsen/logrus"
 )
 
@@ -84,8 +84,8 @@ func main() {
 		MaxTokenLifetime: config.MaxTokenLifetime,
 	}
 
-	targetConfig := target.NewConfig(config.ConfigMapName, config.NSPServiceName, config.NSPServicePort, config.Host, netUtils, apiClientConfig)
-	ambassador, err := target.NewAmbassador(config.Socket, config.Namespace, targetConfig)
+	targetConfig := tap.NewConfig(config.ConfigMapName, config.NSPServiceName, config.NSPServicePort, config.Host, netUtils, apiClientConfig)
+	ambassador, err := tap.NewAmbassador(config.Socket, config.Namespace, targetConfig)
 	if err != nil {
 		logrus.Fatalf("Error creating new ambassador: %v", err)
 	}
