+++ pkg/ambassador/tap/config.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package target
+package tap
 
 import (
 	"github.com/nordix/meridio/pkg/networking"
