+++ pkg/ambassador/tap/stream/stream_test.go
@@ -27,9 +27,9 @@ import (
 	"github.com/golang/protobuf/ptypes/empty"
 	nspAPI "github.com/nordix/meridio/api/nsp/v1"
 	nspMock "github.com/nordix/meridio/api/nsp/v1/mocks"
+	"github.com/nordix/meridio/pkg/ambassador/tap/stream"
+	"github.com/nordix/meridio/pkg/ambassador/tap/types/mocks"
 	"github.com/nordix/meridio/pkg/loadbalancer/types"
-	"github.com/nordix/meridio/pkg/target/stream"
-	"github.com/nordix/meridio/pkg/target/types/mocks"
 	"github.com/stretchr/testify/assert"
 	"google.golang.org/grpc"
 	"google.golang.org/protobuf/types/known/emptypb"
