+++ pkg/ambassador/tap/types/mocks/conduit.go
@@ -10,7 +10,7 @@ import (
 
 	gomock "github.com/golang/mock/gomock"
 	v1 "github.com/nordix/meridio/api/nsp/v1"
-	types "github.com/nordix/meridio/pkg/target/types"
+	types "github.com/nordix/meridio/pkg/ambassador/tap/types"
 )
 
 // MockConduit is a mock of Conduit interface.
